Time: 450 ms (93.65%), Space: 140.3 MB (91.31%) - LeetHub

+++ 2438-range-product-queries-of-powers/2438-range-product-queries-of-powers.cpp
@@ -2,15 +2,17 @@ class Solution {
 public:
     int mod = pow(10, 9) + 7;
     
-    void getPowers(vector<long long>& powers, long long n) {
+    void getPowers(vector<int>& powers, int n) {
         
-        vector<long long> v;
+        vector<int> v;
         
+        // Converting the decimal number 'n' into its binary equivalent.
         while (n > 0) {
             v.push_back(n % 2);
             n /= 2;
         }
         
+        // Bit at i position having 1 is where power of 2 of that 'n' number exists.
         for(int i=0 ; i<v.size() ; i++) {
             if(v[i] == 1) powers.push_back(pow(2, i));
         }
@@ -21,7 +23,7 @@ public:
         
         int q = queries.size();
         vector<int> result(q);
-        vector<long long> powers;
+        vector<int> powers;
         
         getPowers(powers, n);
